+++ src/Index.js
@@ -17,6 +17,7 @@ import {
   getIndexPageOverride,
   skipOverride
 } from './Utils'
+import { Pagination } from './Pagination'
 
 export const DefaultIndexTitle = ({
   schema,
@@ -32,7 +33,7 @@ export const DefaultIndexTitle = ({
   const onCreateClick = R.path(['create', 'onIndexCreate'], actions)
   const onClick = () => onCreateClick({ modelName, path })
   const creatable = isCreatable({ schema, modelName, data, user, customProps })
-  const filterable = isTableFilterable({ schema, modelName })
+  const filterable = isTableFilterable({ schema, modelName, user })
   const currentFilters = R.path(['filter', modelName], tableView)
   const filterOrder = R.path(['filterOrder', modelName], tableView)
   const filtersAreActive = R.path(['filtersAreActive', modelName], tableView)
@@ -48,7 +49,7 @@ export const DefaultIndexTitle = ({
         selectOptions,
         data,
         filterOrder,
-        filterInputs: currentFilters
+        filterInputs: currentFilters, user
       }}/>}
       <div className='float-right'>
         {filterable && <FilterModalButton {...{ modelName, filtersAreActive }} />}
@@ -138,6 +139,13 @@ const DefaultIndex = ({
           }}
         />
       )}
+      {
+        <Pagination {...{
+          schema,
+          modelName,
+          tableView
+        }} />
+      }
     </div>
   )
 }
@@ -181,7 +189,7 @@ const Index = ({
 
   const IndexOverride = getIndexOverride(schema, modelName)
   const IndexComponent = IndexOverride || DefaultIndex
-  
+
   return skipOverride(IndexOverride) ? null : (
     <IndexComponent
       {...{