+++ src/Pagination.js
@@ -1,0 +1,88 @@
+import React from 'react'
+import * as R from 'ramda'
+import { getActions } from './utils/schemaGetters'
+
+const PaginationLink = ({ modelName, changePage, text, updatedPageIndex }) => {
+  return (
+    <li className="page-item">
+      <a className="page-link" href="#" onClick={() => changePage({
+        modelName,
+        updatedPageIndex
+      })}>
+        {text}
+      </a>
+    </li>
+  )
+}
+
+export const Pagination = ({ schema, modelName, tableView }) => {
+  const actions = getActions(schema, modelName)
+  const changePage = R.path(['tableOptions', 'changePage'], actions)
+
+  // current page idx
+  const idx = R.pathOr(0, ['page', modelName], tableView)
+
+  // get index of last hypothetical data point
+  const lastIndex = R.path(['lastIndexPagination', modelName], tableView)
+
+  // get previous & last conditions; 'lastIndex' can be null or '0' value
+  const hasFirst = idx > 1
+  const hasPrev = idx > 0
+  const hasNext = (lastIndex > 0) && (idx < lastIndex)
+  const hasLast = (lastIndex > 0) && (idx < (lastIndex-1))
+
+  // number displayed to user
+  const displayIndex = idx + 1
+
+  // no pagination
+  if (!hasFirst && !hasPrev && !hasNext && !hasLast) {
+    return null
+  }
+
+  return (
+    <nav aria-label="Page navigation example">
+      <ul className="pagination">
+        {
+          hasFirst && <PaginationLink {...{
+            modelName,
+            changePage,
+            text: '«',
+            updatedPageIndex: (0)
+          }} />
+        }
+        {
+          hasPrev && <PaginationLink {...{
+            modelName,
+            changePage,
+            text: '‹',
+            updatedPageIndex: (idx - 1)
+          }} />
+        }
+        {
+          <PaginationLink {...{
+            modelName,
+            changePage,
+            text: `${displayIndex}`,
+            updatedPageIndex: (idx)
+          }} />
+        }
+        {
+          hasNext && <PaginationLink {...{
+            modelName,
+            changePage,
+            text: '›',
+            updatedPageIndex: (idx + 1)
+          }} />
+        }
+        {
+          hasLast && <PaginationLink {...{
+            modelName,
+            changePage,
+            text: '»',
+            updatedPageIndex: (lastIndex)
+          }} />
+        }
+      </ul>
+    </nav>
+  )
+}